% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/randomSort.R
\name{randomSort}
\alias{randomSort}
\title{Sort a vector (title at top of "?" help page)}
\usage{
randomSort(x)
}
\arguments{
\item{x}{a vector ("Arguments" section)}
}
\value{
a sorted version of \code{x} ("Value" section)
}
\description{
Sort a vector into ascending order. ("Description" paragraph)
The algorithm used is the worst one I know.
}
\details{
\code{randomSort} randomly permutes ("Details" section)
            \code{x} until \code{x} is sorted. It requires
            \deqn{(n-1)n!} swaps on average and \deqn{\infty} swaps
            in the worst case.
}
\examples{
v = c(4, 2, 1, 3) # ("Examples" section)
randomSort(x=v)
}

